@@ -22,7 +22,7 @@ use yii\web\Response;
 
 class GetSettingsFormController extends AbstractJobController
 {
-    protected $allowAnonymous = false;
+    protected array|int|bool $allowAnonymous = false;
 
     /**
      * @throws Throwable
