@@ -20,7 +20,7 @@ use yii\web\Response;
 
 class GetJobEditFormController extends AbstractJobController
 {
-    protected $allowAnonymous = false;
+    protected array|int|bool $allowAnonymous = false;
 
     /**
      * @throws Throwable
