@@ -17,7 +17,7 @@ use yii\web\Response;
 
 class GetSyncFromLiltController extends Controller
 {
-    protected $allowAnonymous = false;
+    protected array|int|bool $allowAnonymous = false;
 
     public function actionInvoke(): Response
     {
