@@ -19,7 +19,7 @@ use yii\web\Response;
 
 class PostTranslationPublishController extends AbstractJobController
 {
-    protected $allowAnonymous = false;
+    protected array|int|bool $allowAnonymous = false;
 
     /**
      * @throws Throwable
