@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace lilthq\craftliltplugin\elements\db;
 
 use Craft;
+use craft\db\Query;
 use craft\db\Table;
 use craft\elements\db\ElementQuery;
 use craft\elements\Entry;
@@ -17,9 +18,9 @@ use lilthq\craftliltplugin\elements\Translation;
 
 class TranslationQuery extends ElementQuery
 {
-    public $status;
-    public $dateCreated;
-    public $dateUpdated;
+    public string|array|null $status;
+    public mixed $dateCreated = null;
+    public mixed $dateUpdated = null;
 
     public $jobId;
 
@@ -130,7 +131,7 @@ class TranslationQuery extends ElementQuery
         return parent::beforePrepare();
     }
 
-    public function prepare($builder)
+    public function prepare($builder): Query
     {
         parent::prepare($builder);
 
