@@ -12,7 +12,7 @@ Craft.LiltBaseElementSelectorModal = Craft.BaseElementSelectorModal.extend({
       data.showSiteMenu = this.settings.showSiteMenu ? '1' : '0';
     }
 
-    Craft.postActionRequest('elements/get-modal-body', data,
+    Craft.postActionRequest('element-selector-modals/body', data,
         (response, textStatus) => {
           if (textStatus === 'success') {
             this.$body.html(response.html);
