@@ -34,7 +34,7 @@ window.CraftliltPlugin = {
       }
     });
 
-    $btngroup.prependTo('#header #action-button');
+    $btngroup.prependTo('#header #action-buttons');
 
     Craft.elementIndex.on('selectionChange', function() {
       if (Craft.elementIndex.getSelectedElementIds().length > 0) {
