@@ -60,9 +60,9 @@ CraftliltPlugin.TranslationReview = Garnish.Base.extend({
       this.$modalElementSpinner.hide();
       this.$modalElementBody.append($(response.data));
 
-      $('#lilt-modal-preview-tabs nav ul li a').on('click', function() {
+      $('#lilt-modal-preview-tabs header a').on('click', function() {
         const idToShow = $(this).data('id');
-        $('#lilt-modal-preview-tabs nav ul li a').each(function() {
+        $('#lilt-modal-preview-tabs header a').each(function() {
           const currentId = $(this).data('id');
           if (currentId !== idToShow) {
             $(this).removeClass('sel');
@@ -459,7 +459,7 @@ $(document).ready(function() {
     showStatusMenu: true,
   };
 
-  Craft.postActionRequest('elements/get-modal-body', data,
+  Craft.postActionRequest('element-selector-modals/body', data,
       (response, textStatus) => {
         if (textStatus === 'success') {
           $('#translations-element-index').html(response.html);
