@@ -23,7 +23,7 @@ class JobEdit extends ElementAction
     /**
      * @inheritdoc
      */
-    public function init()
+    public function init(): void
     {
         if ($this->label === null) {
             $this->label = Craft::t('app', 'Edit');
@@ -41,7 +41,7 @@ class JobEdit extends ElementAction
     /**
      * @inheritdoc
      */
-    public function getTriggerHtml()
+    public function getTriggerHtml(): ?string
     {
         $type = Json::encode(static::class);
 
