@@ -12,6 +12,7 @@ namespace lilthq\craftliltplugin\models;
 use Craft;
 use craft\base\Model;
 use craft\helpers\UrlHelper;
+use Exception;
 use lilthq\craftliltplugin\Craftliltplugin;
 use lilthq\craftliltplugin\datetime\DateTime;
 use lilthq\craftliltplugin\records\TranslationRecord;
@@ -99,6 +100,7 @@ class TranslationModel extends Model
             [
                 'elementType' => get_class($element),
                 'sourceId' => $element->getCanonicalId(),
+                'canonicalId' => $element->getCanonicalId(),
                 'draftId' => $element->draftId,
                 'siteId' => $this->sourceSiteId,
             ]
@@ -113,9 +115,14 @@ class TranslationModel extends Model
         );
     }
 
+    /**
+     * @throws Exception
+     */
     public function getLastDeliveryFormatted(): ?string
     {
-        return (new DateTime($this->lastDelivery))->format(Craft::$app->locale->getDateFormat('short', 'php'));
+        return (new DateTime($this->lastDelivery))->format(
+            Craft::$app->locale->getDateFormat('short', 'php')
+        );
     }
 
     public function getDraftEditUrl(): ?string
