@@ -74,7 +74,7 @@ class ElementTranslatableContentApplier
             //TODO: log issue
         }
 
-        $fields = $fieldLayout ? $fieldLayout->getFields() : [];
+        $fields = $fieldLayout ? $fieldLayout->getCustomFields() : [];
 
         foreach ($fields as $field) {
             $fieldData = Craft::$app->fields->getFieldById((int)$field->id);
