@@ -40,12 +40,9 @@ class ElementQueryContentApplier extends AbstractContentApplier implements Appli
         $blockElements = $fieldValue->all();
 
         foreach ($blockElements as $block) {
-            foreach ($block->getFieldLayout()->getFields() as $blockField) {
+            foreach ($block->getFieldLayout()->getCustomFields() as $blockField) {
                 $blockId = $block->getId();
 
-//                $originalSiteBlock = clone $block;
-//                $originalSiteBlock->siteId = $command->getSourceSiteId();
-//                $originalSiteBlockId = $originalSiteBlock->getCanonical()->id;
                 if (
                     !isset(
                         $content[$field->handle][$blockId]['fields'][$blockField->handle]
