@@ -36,9 +36,8 @@ final class ExpectedElementContent
     {
         $matrixContent = self::getExpectedMatrixContent($element, $prefix);
         $neoContent = self::getExpectedNeoContent($element, $prefix, $i18nPrefix);
-        $supertableContent = self::getExpectedSupertableValue($element, $prefix);
 
-        return [
+        $expected = [
             $element->getId() => [
                 'title' => $prefix . 'Some example title',
                 'redactor' => $prefix . '<h1>Here is some header text</h1> Here is some content',
@@ -54,7 +53,6 @@ final class ExpectedElementContent
                     'onLabel' => $i18nPrefix . 'The label text to display beside the lightswitch’s enabled state',
                     'offLabel' => $i18nPrefix . 'The label text to display beside the lightswitch’s disabled state.',
                 ],
-                'supertable' => $supertableContent,
                 'table' => [
                     'columns' => [
                         //TODO: columns heading translation is only done with i18n records
@@ -79,32 +77,64 @@ final class ExpectedElementContent
                     ],
                 ],
                 'neo' => $neoContent,
-                'linkit' => [
-                    'defaultText' => 'Default link text',
-                    'customLabels' =>
-                        [
-                            'fruitstudios\\linkit\\models\\Email' => 'Email address label',
-                            'fruitstudios\\linkit\\models\\Phone' => 'Phone number label',
-                            'fruitstudios\\linkit\\models\\Url' => 'Website url label',
-                        ],
-                    'value' =>
-                        [
-                            'fruitstudios\\linkit\\models\\Email' =>
-                                [
-                                    'value' => 'test@lilt.com',
-                                    'customText' => 'Test linkit text label',
-                                ],
-                        ],
-                ],
-                'colorSwatches' => [
-                    'labels' =>
-                        [
-                            'a5e0af2bdf434712fd71358f5a2415b1' => 'first label',
-                            'e7c9c88325b2a6a2476e2516094b6ba4' => 'second label',
-                            'f13b85cdf5fdd245b03675f94d964946' => 'third label',
-                        ],
-                ],
             ],
+        ];
+
+        if (TEST_SUPERTABLE_PLUGIN) {
+            $supertableContent = self::getExpectedSupertableValue($element, $prefix);
+            $expected[$element->getId()]['supertable'] = $supertableContent;
+        }
+
+        if (TEST_LINKIT_PLUGIN) {
+            $linkitContent = self::getExpectedLinkitContent();
+            $expected[$element->getId()]['linkit'] = $linkitContent;
+        }
+
+        if (TEST_COLOUR_SWATCHES_PLUGIN) {
+            $colourSwatchesContent = self::getExpectedColourSwatchesContent();
+            $expected[$element->getId()]['colorSwatches'] = $colourSwatchesContent;
+        }
+
+        return $expected;
+    }
+
+
+    /**
+     * @throws InvalidFieldException
+     */
+    public static function getExpectedColourSwatchesContent(): array
+    {
+        return [
+            'labels' =>
+                [
+                    'a5e0af2bdf434712fd71358f5a2415b1' => 'first label',
+                    'e7c9c88325b2a6a2476e2516094b6ba4' => 'second label',
+                    'f13b85cdf5fdd245b03675f94d964946' => 'third label',
+                ],
+        ];
+    }
+
+    /**
+     * @throws InvalidFieldException
+     */
+    public static function getExpectedLinkitContent(): array
+    {
+        return [
+            'defaultText' => 'Default link text',
+            'customLabels' =>
+                [
+                    'fruitstudios\\linkit\\models\\Email' => 'Email address label',
+                    'fruitstudios\\linkit\\models\\Phone' => 'Phone number label',
+                    'fruitstudios\\linkit\\models\\Url' => 'Website url label',
+                ],
+            'value' =>
+                [
+                    'fruitstudios\\linkit\\models\\Email' =>
+                        [
+                            'value' => 'test@lilt.com',
+                            'customText' => 'Test linkit text label',
+                        ],
+                ],
         ];
     }
 
@@ -175,7 +205,7 @@ final class ExpectedElementContent
         $firstBlockId = $neoFieldValue->type('firstBlockType')->one()->id;
         $secondBlockId = $neoFieldValue->type('secondBlockType')->one()->id;
 
-        return [
+        $content = [
             $firstBlockId => [
                 'fields' => [
                     'redactor' => $prefix . 'firstBlockType - redactor - Here is value of field',
@@ -209,5 +239,11 @@ final class ExpectedElementContent
                 ],
             ],
         ];
+
+        if (TEST_SUPERTABLE_PLUGIN) {
+            $content[$secondBlockId]['fields']['supertable'] = [];
+        }
+
+        return $content;
     }
 }
