@@ -108,7 +108,6 @@ class GetJobEditFormControllerCest
         ]);
 
         $controller = $this->getController();
-        $controller->setView(new ViewWrapper());
 
         $I->expectSettingsGetRequest(
             '/api/v1.0/settings',
@@ -123,9 +122,14 @@ class GetJobEditFormControllerCest
 
         $response = $controller->actionInvoke((string)$job->id);
 
-        $expected = $this->getExpected($job);
+        $behavior = $response->getBehavior('template');
+        $actual = [
+            'variables' => $behavior->variables ?? [],
+            'template' => $behavior->template ?? '',
+            'templateMode' => $behavior->templateMode ?? '',
+        ];
 
-        $actual = json_decode($response->data, true, 512, 4194304);
+        $expected = $this->getExpected($job);
 
         Assert::assertEquals($expected['template'], $actual['template']);
         Assert::assertEquals(
@@ -149,6 +153,9 @@ class GetJobEditFormControllerCest
             $actual['variables']['authorOptionCriteria']
         );
         Assert::assertEquals($expected['variables']['crumbs'], $actual['variables']['crumbs']);
+
+        #convert to array
+        $actual['variables']['element'] = json_decode(json_encode($actual['variables']['element']), true);
 
         foreach ($expected['variables']['element'] as $expectedKey => $expectedValue) {
             Assert::assertArrayHasKey($expectedKey, $actual['variables']['element']);
@@ -324,11 +331,11 @@ class GetJobEditFormControllerCest
                 'crumbs' => [
                     0 => [
                         'label' => 'Lilt Plugin',
-                        'url' => 'http://$PRIMARY_SITE_URL/index.php?p=admin/admin/craft-lilt-plugin',
+                        'url' => 'http://$PRIMARY_SITE_URL/index.php?p=admin/admin/craft-lilt-plugin&site=default',
                     ],
                     1 => [
                         'label' => 'Jobs',
-                        'url' => 'http://$PRIMARY_SITE_URL/index.php?p=admin/admin/craft-lilt-plugin/jobs',
+                        'url' => 'http://$PRIMARY_SITE_URL/index.php?p=admin/admin/craft-lilt-plugin/jobs&site=default',
                     ],
                 ],
             ],
